[feat/val-client] move network code to simulations/
